requirements.txt
chardet==5.2.0
langchain==1.0.5
langchain_community==0.4.1
langchain_core==1.0.4
langchain_openai==1.0.2
langchain_text_splitters==1.0.0
langgraph==1.0.3
pycparser==2.23
python-dotenv==1.2.1
tavily==1.1.0
typing_extensions==4.15.0
fastapi>=0.115.5,<0.122.0
uvicorn[standard]>=0.32.0,<0.39.0
sqlalchemy>=2.0.36,<2.1.0
aiomysql>=0.2.0,<0.4.0
aiosqlite>=0.20.0,<0.21.0
pydantic-settings>=2.10.1,<3.0.0
alembic==1.14.0
passlib[bcrypt]==1.7.4
httpx>=0.27.2,<0.28.0
pytest>=8.3.3,<9.0.0
pytest-asyncio>=0.24.0,<0.25.0
